v2.2

+++ celadorecs/recslib.py
@@ -208,7 +208,7 @@ class RecommenderSystem():
         num_cols.drop(self.id_vars, axis=1, inplace=True, errors='ignore')
         if cat_cols.shape[1] > 0:
             cat_cols = cat_cols.astype('category')    
-            one_hot = pd.get_dummies(cat_cols, drop_first=True)
+            one_hot = pd.get_dummies(cat_cols, drop_first=True).astype(np.int8)
             to_drop = list(filter(lambda s:s[-3:]=='_no' \
                                   or s[-4:]=='_0.0' \
                                   or s[-2:]=='_0',
